<ExerciseMain 버튼에 이미지 추가>
