+++ src/Main.java
@@ -1,14 +1,10 @@
-import food.*;
-import exercise.*;
+import cal.food.FoodMain;
 
-// Enter를 누르세요. 그러면 코드 내에서 공백 문자를 확인할 수 있습니다.
 public class Main {
     public static void main(String[] args) {
-        // Shift+F9을(를) 눌러 코드 디버그를 시작합니다. 중단점을 하나 설정해 드렸습니다.
-        // 중단점을 더 추가하려면 언제든지 Ctrl+F8을(를) 누르세요
 
-        new ExerciseMain();
+        //new ExerciseMain();
+        new FoodMain();
 
-        //new FoodMain();
     }
 }
