+++ src/cal/Images/출처.txt
@@ -1,0 +1,11 @@
+<a href="https://www.flaticon.com/free-icons/people" title="people icons">People icons created by Bharat Icons - Flaticon</a> 걷기 아이콘
+<a href="https://www.flaticon.com/free-icons/run" title="run icons">Run icons created by Freepik - Flaticon</a> 달리기 아이콘
+<a href="https://www.flaticon.com/free-icons/stair" title="stair icons">Stair icons created by Freepik - Flaticon</a> 계단오르내리기 아이콘
+<a href="https://www.flaticon.com/free-icons/skipping-rope" title="skipping rope icons">Skipping rope icons created by Iconriver - Flaticon</a> 줄넘기 아이콘
+<a href="https://www.flaticon.com/free-icons/enjoy" title="enjoy icons">Enjoy icons created by Mayor Icons - Flaticon</a> 댄스 아이콘
+<a href="https://www.flaticon.com/free-icons/bycicle" title="bycicle icons">Bycicle icons created by iconographics - Flaticon</a> 자전거 아이콘
+<a href="https://www.flaticon.com/free-icons/sit-up" title="sit up icons">Sit up icons created by ARIF MIFTAHUL ANWAR - Flaticon</a> 윗몸일으키기 아이콘
+<a href="https://www.flaticon.com/free-icons/push-ups" title="push ups icons">Push ups icons created by Leremy - Flaticon</a> 팔굽혀펴기 아이콘
+<a href="https://www.flaticon.com/free-icons/squatting" title="squatting icons">Squatting icons created by Leremy - Flaticon</a> 스쿼트 아이콘
+<a href="https://www.flaticon.com/free-icons/swimming" title="swimming icons">Swimming icons created by Freepik - Flaticon</a> 수영 아이콘
+
